Add OS ex3
MapReduce is used to parallelize tasks of a specific structure. 
The map function is applied to each input element, producing a sequence of intermediary elements.
The intermediary elements are sorted into new sequences.
The reduce function is applied to each of the sorted sequences of intermediary elements, producing a sequence of output elements.