Train a combined model
Use a VGG-FACE encoder to extract features from both the target images and source images into shared feature space. 
Build a combined classifier, a combined discriminator, and a combined model of the classifier and discriminator.
Adversarial learning - train a combined model when the discriminator is non-trainable.
Use the combined classifier to classify target images.